move network memory
