+++ xlive/Blam/Engine/simulation/simulation_entity_database.cpp
@@ -4,7 +4,7 @@
 #include "simulation.h"
 #include "simulation_queue_entities.h"
 
-#include "networking/memory/networking_memory.h"
+#include "networking/network_memory.h"
 
 
 c_simulation_entity_database* simulation_get_entity_database()
