+++ xlive/Blam/Engine/simulation/simulation_queue.cpp
@@ -3,7 +3,7 @@
 
 #include "simulation_queue_global_events.h"
 
-#include "networking/memory/networking_memory.h"
+#include "networking/network_memory.h"
 
 void c_simulation_queue::allocate(int32 size, s_simulation_queue_element** out_allocated_elem)
 {
